refactor: rename controller to route
