+++ nodejs/src/app.ts
@@ -1,9 +1,9 @@
 import express from 'express';
 import cors from 'cors';
-import { router } from './controller';
+import { route } from './route';
 
 export const app = express();
 
 app.use(cors());
 app.use(express.json());
-app.use('/', router);
+app.use('/', route);
