--- nodejs/src/controller/index.ts
@@ -1,2 +1,0 @@
-export * from './controller.interface';
-export * from './controller';
