+++ nodejs/src/route/index.ts
@@ -1,0 +1,2 @@
+export * from './route.interface';
+export * from './route';
