+++ nodejs/src/route/route.ts
@@ -1,11 +1,11 @@
 import express from 'express';
 import logger from '../logger';
 import { generateGameSession } from '../services/bot/bot.service';
-import { ResponderBaseChannelConfig } from './controller.interface';
+import { ResponderBaseChannelConfig } from './route.interface';
 
-export const router = express.Router();
+export const route = express.Router();
 
-router.post('/open', (async (req, res) => {
+route.post('/open', (async (req, res) => {
   const reqBody = req.body as Partial<ResponderBaseChannelConfig>;
   const { address, port, host } = reqBody;
   try {
@@ -28,4 +28,4 @@ router.post('/open', (async (req, res) => {
   }
 }) as express.RequestHandler);
 
-export default router;
+export default route;
